+++ .gitignore
@@ -5,3 +5,5 @@
 *.zip
 *.pyc
 *.csv
+data/.DS_Store
+.DS_Store
